+++ src/backends/sentry_backend_breakpad.cpp
@@ -186,14 +186,14 @@ breakpad_backend_callback(const google_breakpad::MinidumpDescriptor &descriptor,
                 envelope, dump_path, "attachment");
             if (item) {
                 sentry__envelope_item_set_header(item, "attachment_type",
-                    sentry_value_new_string("event.minidump"));
+                    sentry_value_new_string(SENTRY_ATTACHMENT_TYPE_MINIDUMP));
 
                 sentry__envelope_item_set_header(item, "filename",
                     sentry_value_new_string(sentry__path_filename(dump_path)));
             } else if (options->enable_large_attachments) {
                 sentry_attachment_t tmp = {};
                 tmp.path = dump_path;
-                tmp.type = MINIDUMP;
+                tmp.type = (char *)SENTRY_ATTACHMENT_TYPE_MINIDUMP;
                 if (!sentry__cache_attachment_ref(
                         envelope, &tmp, options->run->cache_path, nullptr)) {
                     SENTRY_SIGNAL_SAFE_LOG(