+++ src/backends/sentry_backend_crashpad.cpp
@@ -517,8 +517,8 @@ report_to_envelope(const crashpad::CrashReportDatabase::Report &report,
             } else if (strcmp(filename, "__sentry-breadcrumb2") == 0) {
                 breadcrumbs2 = read_msgpack_file(path);
             } else {
-                sentry__attachments_add_path(&attachments,
-                    sentry__path_clone(path), ATTACHMENT, nullptr);
+                sentry__attachments_add_path(
+                    &attachments, sentry__path_clone(path), nullptr, nullptr);
             }
         }
         sentry__pathiter_free(iter);
@@ -537,8 +537,8 @@ report_to_envelope(const crashpad::CrashReportDatabase::Report &report,
         sentry_value_set_by_key(event, "breadcrumbs",
             sentry__value_merge_breadcrumbs(
                 breadcrumbs1, breadcrumbs2, options->max_breadcrumbs));
-        sentry__attachments_add_path(
-            &attachments, minidump_path, MINIDUMP, nullptr);
+        sentry__attachments_add_path(&attachments, minidump_path,
+            SENTRY_ATTACHMENT_TYPE_MINIDUMP, nullptr);
 
         if (sentry__envelope_add_event(envelope, event)) {
             sentry__envelope_add_attachments(envelope, attachments, options);